Fix parsing of Scala classes containing symbols
Scala classes are allowed to have symbols like +-~/: in them.
When compiled, these become Java class names by replacing the symbol
with a $ and a word. + becomes $plus. The code for parsing class names
for inner classes didn't handle these cases and would throw errors
saying that classes didn't exist when sonar was run. This fixes the
parsing.