QA: fix test of version after change of default maven scanner behavior
Maven scanner behavior is to know rely on which java version was resolved by maven.
As such, if the version is unset, it is likely to be resolved to java 5 and be set for the analysis.
This changes the way we tested unset java version to use scanner for this case.