Rule S1188: Add separate parameter for lambdas
The length of a body of a lambda expression and an anonymous class should be
treated differently: Anonymous classes may contain more than a single method
and should be permitted to have a longer body. Lambdas on the other hand
should be short and concise and thus require a different setting.