S2599 FP vertx asyncAssertSuccess, asyncAssertFailure
Vertx' TestContext has four methods for asynchonous assertions:
Two without arguments:
* asyncAssertSuccess()
* asyncAssertFailure()
Two with Handler arguments:
* asyncAssertSuccess(Handler<T> resultHandler)
* asyncAssertFailure(Handler<Throwable> causeHandler)

All four either resolve the Async or fail the test.

But the methods without arguments are the end of the test because no other code is executed
whereas the methods with Handler arguments continue with the handler code.

The methods without arguments are intended for cases where checking success or failure is sufficient,
therefore they should count as a S2599 assertion.

The methods with arguments are intended for cases where more testing is needed:
Either testing the result (or the Throwable of the failure) or by executing some
other operation. Therefore the methods with arguments should not count as a S2599
assertion.

Documentation: https://vertx.io/docs/vertx-unit/java/#_asynchronous_assertions