False positive fix on Nested JUnit 5 classes on S3577
Rule S3577 checks a name of the test to comply with naming scheme. It
checks for a name if a class contains a test method. JUnit 5 introduced
nested classes that should be excluded from this rule as they do not
need to be searched by surefire/failsafe or other tests executors.

In this commit check has been changed to take acount only top level
classes that contain tests and/or nested classes.