Refactoring to avoid duplicated code for flags and error handling
Make the result of parsing a regex contain information about which
flags the regex was created with.

Use this information in the RegexBaseVisitor to set the initial
flags, removing any need to call setFlags, which is now removed.

Also move the logic of checking whether the regex contains syntax
errors before visiting anything into the RegexBaseVisitor and the
methods before and after to it, so now rules should all be
implementable with only a suitable definition of a visitor class
and then `new MyVisitor(possibleArguments).visit(regexForLiterals);`.