Fix quality flaw: simplify regex (#3546)
* Regex: replace alternation with character class

* Simplify regex in Javadoc helper class

The regex itself is subject to catastrophic backtracking, when the input is like "@aaa\n", with thousands of "a". In practice, we already process the string that will be matched by the regex, so I don't think we are really subject to a problem here. Still, the regex can be simplified because there is no point in being that precise to detect "Block tag", a line starting with "@" is already enough.

* Reduce complexity of Regex in HardcodedIpCheck