SONARJAVA-4098 Update S1612: Short lambdas are considered compliant (#4238)
After this change, no issue will be raised when a lambda is shorter than its equivalent method reference.

To determine whether a lambda is shorter than its replacement, we make an approximation and only measure the size of single-line lambdas. The reason is that accurately getting the original code's length is hard when it spans multiple line. This means we will report multi-line lambdas that have an equivalent method reference and are shorter than it, but this is very rare in practice and is the status quo.