SONARJAVA-4420 Fix IndexOutOfBoundsException on switch expression
Reduced approximation involved in cleaning up (clearing the stack). Instead, we now try to take off only what we should from the stack. Also, exceptions that are the cause for a path to be walked are not stored on the stack anymore but instead have a dedicated field in the ProgramState."