SONARJAVA-4420 Fixed exception due to stack pollution (#4386)
Previously, we regularly nuked the stack. Since 82ee22d49f959284e72470110998c507129ce948 we don't do this anymore and instead try to handle the stack in a bit more coordinated manner. However, for logical AND and OR operations, this exposed some stack pollution. This commit works around a prior hack to avoid this stack pollution.