SONARJAVA-4988: Use SonarLintCache component and make it accessible to custom rules via the caching APIs (#4792)
To enable DBD support in SonarLint for Java in VSCode, DBD needs to be able to access the intermediate representation (IR) files it generates for the Java code under analysis.
This IR is generated by custom rules for sonar-java which are provided by DBD, and usually it is stored in the file system. However, no file system is available in a SonarLint context.

Hence, the IR needs to be transferred in memory. For DBD Python analysis, this has been achieved by utilizing a cache context.
I.e. a component SonarLintCache is injected into the Python analyzer frontend, a CacheContext is constructed from it, and DBD’s custom rules store the IR in this cache.
Then, when the DBD plugin is executed, it can retrieve the IR from the cache.

This PR applies the same change to sonar-java.

---

* SONARJAVA-4988: Expose SonarProduct on ModuleScannerContext

DBD custom rules need this information to turn off saving IR to the
filesystem in a SonarLint context

* SONARJAVA-4988: Always provide CacheContext if SonarLintCache is available

* SONARJAVA-4988: CacheContexts based on SonarLintCache should not report as a proper cache

* SONARJAVA-4988: Permit sensor execution ordering using @DependedUpon annotations