SONARJAVA-4881 Clearly separate Spring5 and Spring6 samples
We now clearly separate the Spring 5 and Spring 6 sample files because the issue raised in the two cases will be different.
In Spring 5, we raise for any non-public method annotated with the relevant annotation.
In Spring 6, only for private methods.
We also check that when Spring dependencies are not part of the classpath, we do not raise.

Not that in the Spring 6 case, we use fully qualified annotation name. This is because we use a trick for testing by providing a fake classpath. This is to make the check class think we are using Spring 6. But since the Spring 6 jar is not really available for the project so the annotation would not be resolved.