SONARJAVA-4376 Fix S2129 in the case of missing semantic
When semantic is missing for the argument of String constructor for instance, the parser takes a guess at which constructor it is and this may cause some false positives.
The fix in this PR, skips the case where we detect an unknown in the argument types.
The second sample is coming from SONARJAVA-5388